+++ scrabble_helper/main.py
@@ -2,7 +2,7 @@ from scrabble_helper.io import advise_from_json
 
 
 def main():
-    advise_from_json("test")
+    advise_from_json("game")
 
 
 if __name__ == "__main__":